% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_city.R
\name{plot_city}
\alias{plot_city}
\title{Plot for selected city}
\usage{
plot_city(data, city)
}
\arguments{
\item{data}{A data frame containing a \code{city} variable}

\item{city}{Choose city to plot}
}
\description{
Plot total confirmed cases for selected city
}
\examples{
 
\dontrun{ 
plot_city(cov_china, "Ankang")
}

}
